Generalize to site or sponsor data
Some input files will have city/country data (generally the site data),
so here we allow the user to specify the file type.

Add in harcoded table generation for the processed files.

When querying, first group by name or name/city/country so we do not
have to make duplicate queries to ror (when not using the cache).