Add fixes to handle data from different registries
The different registries have different numbers of columns so make sure
we reindex on any merge
We might use a mapping dictionary from one registiry site/sponsor and
apply it on a registry with more or less info. Assert that the merge has
happened correctly.

As a partial fix to #12 we remove double quotes from names, but if there
is a special character, as a last step try adding double quotes.