Integrate PubMed and OpenAlex
Additionally get publication date, accession #s and extract any
registry ids from the abstract.

For openalex, instead of using a text file of pmids, use the csv file
with extra pubmed data. Also make sure we add a row to the output either
if there was no entry in openalex, or the author had no resolved
institutions.

Data is unique by: pmid, author, and institution as a paper can have a
first/last author, and each author can have multiple institutions. A
current limitation of this method is that openalex may have an
institution associated with an author that was not involved in the given
trial (not in the raw affiliation string).